refactor: Split AllLocationsMap.tsx into focused frontend modules (R14)
Extract popup sizing, popup HTML, Leaflet setup, and shared types so the
React component stays a thin lifecycle shell; rebuild committed bundle.
