+++ docs/development.md
@@ -75,7 +75,7 @@ All four **Map view** modes use the same Streamlit custom component (`explorer/c
 | Prep | `app_prep_map_ui.py` (orchestrator) + `app_prep_map_leaflet_modes.py` | Build GeoJSON + `revision`, banner/legend HTML, viewport extras; session LRU via `app_prep_map_leaflet_caches` |
 | Payload | `explorer/core/*_locations_geojson.py` | Structured `popup_v1` (and mode variants) in feature properties — not per-pin HTML |
 | Embed | `render_all_locations_map_component` in component `__init__.py` | `declare_component` + committed `frontend/build` iframe |
-| Client | `frontend/src/AllLocationsMap.tsx` | Leaflet map, MarkerCluster, popup templates (`AllLocationsMapPopup.css`) |
+| Client | `frontend/src/AllLocationsMap.tsx` (+ `AllLocationsMapLeaflet.ts`, `AllLocationsMapPopupHtml.ts`, `AllLocationsMapPopupSizing.ts`, `allLocationsMapTypes.ts`) | Leaflet map, MarkerCluster, popup templates (`AllLocationsMapPopup.css`) |
 | Export | `explorer/presentation/leaflet_map_html_export.py` | Standalone HTML (CDN Leaflet) from cached recipe (`LEAFLET_EXPORT_*` keys) |
 
 **Historical performance notes** (#222): [`issue-222-plain-summary.md`](explorer/issue-222-plain-summary.md) (plain language), [`issue-222-section-8-baseline.md`](explorer/issue-222-section-8-baseline.md) (tables + re-run), [`issue-222-section-8-prior-art.md`](explorer/issue-222-section-8-prior-art.md) (Folium-era context). Re-run: `./scripts/run_post_leaflet_perf_baseline.sh`.
