+++ explorer/components/all_locations_map/frontend/src/AllLocationsMap.tsx
@@ -5,1216 +5,38 @@ import "leaflet.markercluster/dist/MarkerCluster.css";
 import "leaflet.markercluster/dist/MarkerCluster.Default.css";
 import "leaflet.markercluster/dist/leaflet.markercluster.js";
 import "./AllLocationsMapPopup.css";
-import {
-  parseViewportV1,
-  type ViewportV1GoToGps,
-} from "./mapComponentParsers";
+import { parseViewportV1 } from "./mapComponentParsers";
 import {
   ComponentProps,
   Streamlit,
   withStreamlitConnection,
 } from "streamlit-component-lib";
+import type { MapArgs } from "./allLocationsMapTypes";
+import {
+  applyAllLocationsViewport,
+  applyBasemapToMap,
+  applyGoToGpsViewportCamera,
+  injectHeadFragments,
+  markerClusterGroupOptionsWithOptionalIconStyle,
+  mergeClusterPayload,
+  parseCirclePin,
+  resolvedCircleStyles,
+  resolvedCircleStylesFromPinPayload,
+  syncGoToGpsMarker,
+  syncZoomLevelDebugOverlay,
+} from "./AllLocationsMapLeaflet";
+import { popupHtmlFromFeatureProps } from "./AllLocationsMapPopupHtml";
+import {
+  POPUP_BIND_OPTIONS,
+  maybePanPopupIntoView,
+  scheduleShrinkPebirdLeafletPopups,
+  shrinkPebirdLeafletPopups,
+} from "./AllLocationsMapPopupSizing";
 
 declare global {
   interface Window {
     /** Playwright E2E hook: Leaflet map in this component iframe (not on window by default). */
     __pebirdLeafletMap?: L.Map;
-  }
-}
-
-interface ClusterOptionsPayload {
-  enabled?: boolean;
-  max_cluster_radius?: number;
-  disable_clustering_at_zoom?: number;
-  spiderfy_on_max_zoom?: boolean;
-  remove_outside_visible_bounds?: boolean;
-}
-
-interface CircleMarkerStylePayload {
-  fill_hex?: string;
-  stroke_hex?: string;
-  radius_px?: number;
-  stroke_weight?: number;
-  fill_opacity?: number;
-}
-
-/** Folium ``_marker_cluster_icon_create_function_from_scheme`` parity — from :func:`all_locations_cluster_icon_style_payload`. */
-interface ClusterIconStylePayload {
-  fills_rgba: string[];
-  borders_rgba: string[];
-  halos_rgba: string[];
-  border_width_px: number;
-  halo_spread_px: number;
-}
-
-/** Folium `MAP_POPUP_MAX_WIDTH_PX` (`explorer/app/streamlit/defaults.py`). */
-const POPUP_MAX_WIDTH_PX = 420;
-
-/** Leaflet runs ``autoPan`` inside ``popup.update()`` — stacked updates caused large vertical pans. Disabled globally; ``maybePanPopupIntoView`` pans once when needed after layout settles. */
-const POPUP_BIND_OPTIONS: L.PopupOptions = {
-  maxWidth: POPUP_MAX_WIDTH_PX,
-  autoPan: false,
-};
-
-/** Folium ``_apply_go_to_gps_pin_view`` popup body (`explorer/core/map_overlay_visit_map.py`). */
-const GO_TO_GPS_POPUP_HTML =
-  "<div style='font-size:13px'><strong>Temporary GPS marker</strong></div>";
-
-/** Default gap below location title row — matches ``build_location_popup_html(..., location_heading_margin_px=4)``. */
-const POPUP_LOCATION_HEADING_MARGIN_PX = 4;
-
-/** Extra px on shrink width — ``scrollWidth`` can sit slightly under painted text (subpixel / fonts / padding). */
-const POPUP_SHRINK_WIDTH_BUFFER_PX = 48;
-
-/** Never apply a narrower content box than this — guards bad measures / font glitches. */
-const POPUP_SHRINK_MIN_CONTENT_WIDTH_PX = 140;
-
-/** Rows included when measuring intrinsic popup width (all map modes). */
-const POPUP_WIDE_MEASURE_SELECTOR =
-  ".pebird-map-popup__heading-row, .pebird-map-popup__visit-dates a, .pebird-map-popup__visit-list-inner a, " +
-  ".pebird-map-popup__summary-line, " +
-  ".pebird-map-popup__species-line, .pebird-map-popup__species-line a, .pebird-map-popup__obs-line, " +
-  ".pebird-map-popup__species-seen > summary, " +
-  ".pebird-map-popup__all-visits > summary";
-
-const POPUP_LOCATION_HEADING_SELECTOR =
-  "a.pebird-map-popup__location-heading, span.pebird-map-popup__location-heading";
-
-/** ``.pebird-map-popup`` horizontal padding (6 + 14) + ``__heading-row`` close-button gutter (~2.25rem). */
-const POPUP_HEADING_TEXT_INSET_PX = 56;
-
-/** Intrinsic single-line width of location titles (off-DOM — matches painted bold popup font). */
-function measureLocationHeadingWidthPx(inner: HTMLElement): number {
-  let w = 0;
-  inner.querySelectorAll(POPUP_LOCATION_HEADING_SELECTOR).forEach((el) => {
-    const he = el as HTMLElement;
-    const cs = window.getComputedStyle(he);
-    const clone = he.cloneNode(true) as HTMLElement;
-    clone.style.cssText =
-      "position:absolute;left:-9999px;top:0;visibility:hidden;white-space:nowrap;width:max-content;max-width:none;";
-    clone.style.fontFamily = cs.fontFamily;
-    clone.style.fontSize = cs.fontSize;
-    clone.style.fontWeight = cs.fontWeight;
-    clone.style.letterSpacing = cs.letterSpacing;
-    document.body.appendChild(clone);
-    w = Math.max(w, clone.scrollWidth, clone.getBoundingClientRect().width);
-    document.body.removeChild(clone);
-  });
-  return Math.ceil(Math.max(w, 1));
-}
-
-function headingTextBudgetPx(contentWidthPx: number): number {
-  return Math.max(1, contentWidthPx - POPUP_HEADING_TEXT_INSET_PX);
-}
-
-function committedPopupContentWidthPx(content: HTMLElement): number {
-  const fromData = Number.parseInt(content.dataset.pebirdShrinkTarget ?? "", 10);
-  if (Number.isFinite(fromData) && fromData > 0) {
-    return fromData;
-  }
-  return Math.ceil(content.getBoundingClientRect().width);
-}
-
-/** Drop a too-narrow width commit when the single-line title needs a wider card (or the max cap). */
-function popupWidthCommitTooNarrowForHeading(
-  inner: HTMLElement,
-  content: HTMLElement,
-  cap: number,
-): boolean {
-  const committed = committedPopupContentWidthPx(content);
-  const headingW = measureLocationHeadingWidthPx(inner);
-  return headingW > headingTextBudgetPx(committed) && committed < cap;
-}
-
-/** Max of inner and wide text rows while inner is ``max-content`` for measure. */
-function measurePebirdPopupInnerWidthPx(inner: HTMLElement): number {
-  void inner.offsetWidth;
-  let w = Math.max(inner.scrollWidth, inner.getBoundingClientRect().width);
-  w = Math.max(w, measureLocationHeadingWidthPx(inner));
-  const wideEls = inner.querySelectorAll(POPUP_WIDE_MEASURE_SELECTOR);
-  wideEls.forEach((el) => {
-    const he = el as HTMLElement;
-    w = Math.max(w, he.scrollWidth, he.getBoundingClientRect().width);
-  });
-  return Math.ceil(Math.max(w, 1));
-}
-
-/** Shrink-wrap Leaflet popup width to ``.pebird-map-popup`` intrinsic width (same idea as Folium ``map_popup_width_fix_script``; this iframe runs TS only).
-
-Uses map pixel width (not ``window``) for cap. Parents use ``cap`` px during measure so ``width:100%`` rows do not collapse.
-After width changes, callers invoke ``popup.update()`` to keep the tip on the marker (#145).
-*/
-function capPopupInnerWidthPxForMap(map: L.Map): number {
-  const px = Math.max(1, map.getSize().x);
-  return Math.min(POPUP_MAX_WIDTH_PX, Math.max(80, px - 24));
-}
-
-/** Once width is applied for a given map cap, skip remeasuring — avoids a second visible resize when
- * iframe/RFO bumps run later at ~50–500ms. Remeasure only when ``cap`` changes (map container width). */
-const POPUP_WIDTH_COMMIT_ATTR = "data-pebird-popup-width-commit";
-const POPUP_WIDTH_CAP_ATTR = "data-pebird-shrink-applied-cap";
-
-/** Shrink-wrap ``.leaflet-popup-content`` to intrinsic width; commit per (popup, map cap) to avoid repeat work. */
-function shrinkPebirdLeafletPopups(map: L.Map): void {
-  const pops = document.querySelectorAll(".leaflet-popup-pane .leaflet-popup");
-  const cap = capPopupInnerWidthPxForMap(map);
-  const capStr = `${cap}`;
-  pops.forEach((pop) => {
-    const content = pop.querySelector(".leaflet-popup-content") as HTMLElement | null;
-    const wrap = pop.querySelector(".leaflet-popup-content-wrapper") as HTMLElement | null;
-    const inner = pop.querySelector(".pebird-map-popup") as HTMLElement | null;
-    if (!content || !wrap || !inner) {
-      return;
-    }
-    if (content.getAttribute(POPUP_WIDTH_COMMIT_ATTR) === "1" && content.getAttribute(POPUP_WIDTH_CAP_ATTR) === capStr) {
-      if (popupWidthCommitTooNarrowForHeading(inner, content, cap)) {
-        content.removeAttribute(POPUP_WIDTH_COMMIT_ATTR);
-      } else {
-        return;
-      }
-    }
-
-    /* Shrink-to-fit cycle: inner has max-width:100% of .leaflet-popup-content while that node uses
-     * width:fit-content — cyclic percentage resolves tiny, so scrollWidth was ~min-content  .
-     * Size inner to max-content for measurement only, then restore so final layout still respects cap.
-     *
-     * Never strip content/wrapper width **without** substituting ``cap``: descendants such as
-     * ``.pebird-map-popup__heading-row { width:100% }`` lose their percentage base and collapse to a
-     * min-content column — ``scrollWidth`` then matches a character-wide strip. */
-    inner.style.setProperty("max-width", "none", "important");
-    inner.style.setProperty("width", "max-content", "important");
-    content.style.setProperty("width", `${cap}px`, "important");
-    content.style.setProperty("max-width", `${cap}px`, "important");
-    wrap.style.setProperty("width", `${cap}px`, "important");
-    wrap.style.setProperty("max-width", `${cap}px`, "important");
-    content.style.removeProperty("white-space");
-    const innerPx = measurePebirdPopupInnerWidthPx(inner);
-    inner.style.removeProperty("max-width");
-    inner.style.removeProperty("width");
-    const target = Math.max(
-      POPUP_SHRINK_MIN_CONTENT_WIDTH_PX,
-      Math.min(innerPx + POPUP_SHRINK_WIDTH_BUFFER_PX, cap),
-    );
-    const nextTargetStr = `${target}`;
-    const prevTargetStr = content.dataset.pebirdShrinkTarget ?? "";
-    const changed = prevTargetStr !== nextTargetStr;
-
-    if (changed) {
-      content.dataset.pebirdShrinkTarget = nextTargetStr;
-      content.style.setProperty("width", `${target}px`, "important");
-      content.style.setProperty("max-width", `${cap}px`, "important");
-      wrap.style.setProperty("width", `${target}px`, "important");
-      wrap.style.setProperty("max-width", `${cap}px`, "important");
-    }
-    content.setAttribute(POPUP_WIDTH_COMMIT_ATTR, "1");
-    content.setAttribute(POPUP_WIDTH_CAP_ATTR, capStr);
-  });
-}
-
-/** Margin inside the map container when deciding if the popup clips (replaces Leaflet ``autoPan``). */
-const POPUP_VIEWPORT_PAD_PX = 12;
-
-/** Pan the map only when the open popup’s bounding box exceeds the container inset — minimal correction, no Leaflet ``autoPan`` stack. */
-function maybePanPopupIntoView(map: L.Map, popup: L.Popup): void {
-  const mapEl = map.getContainer();
-  const popupEl = popup.getElement();
-  if (!popupEl) {
-    return;
-  }
-  const mr = mapEl.getBoundingClientRect();
-  const pr = popupEl.getBoundingClientRect();
-  const insetLeft = mr.left + POPUP_VIEWPORT_PAD_PX;
-  const insetTop = mr.top + POPUP_VIEWPORT_PAD_PX;
-  const insetRight = mr.right - POPUP_VIEWPORT_PAD_PX;
-  const insetBottom = mr.bottom - POPUP_VIEWPORT_PAD_PX;
-
-  const overflowLeft = insetLeft - pr.left;
-  const overflowRight = pr.right - insetRight;
-  const overflowTop = insetTop - pr.top;
-  const overflowBottom = pr.bottom - insetBottom;
-
-  let dx = 0;
-  let dy = 0;
-  if (overflowLeft > 0 && overflowRight <= 0) {
-    dx = -overflowLeft;
-  } else if (overflowRight > 0 && overflowLeft <= 0) {
-    dx = overflowRight;
-  } else if (overflowLeft > 0 && overflowRight > 0) {
-    dx = (overflowRight - overflowLeft) / 2;
-  }
-  if (overflowTop > 0 && overflowBottom <= 0) {
-    dy = -overflowTop;
-  } else if (overflowBottom > 0 && overflowTop <= 0) {
-    dy = overflowBottom;
-  } else if (overflowTop > 0 && overflowBottom > 0) {
-    dy = (overflowBottom - overflowTop) / 2;
-  }
-
-  if (dx !== 0 || dy !== 0) {
-    map.panBy(L.point(dx, dy), { animate: false });
-  }
-}
-
-function scheduleShrinkPebirdLeafletPopups(map: L.Map, popup?: L.Popup): void {
-  /** Wait for web fonts before measuring — fallback metrics used to lock ``data-pebird-popup-width-commit``
-   * produced cards that were tens of px too narrow (lines broke after commas, visit rows split date/time).
-   * One double-rAF after ``fonts.ready`` keeps a single width commit with no follow-up resize. */
-  const finalize = () => {
-    shrinkPebirdLeafletPopups(map);
-    if (popup && typeof popup.update === "function") {
-      popup.update();
-      maybePanPopupIntoView(map, popup);
-    }
-  };
-  const runAfterLayout = () => {
-    requestAnimationFrame(() => {
-      requestAnimationFrame(finalize);
-    });
-  };
-  try {
-    const ready = document.fonts?.ready;
-    if (ready && typeof ready.then === "function") {
-      void ready.then(runAfterLayout).catch(runAfterLayout);
-    } else {
-      runAfterLayout();
-    }
-  } catch {
-    runAfterLayout();
-  }
-}
-
-/** Structured popup from Python (`explorer/core/all_locations_geojson.py`) — extend for classic parity. */
-interface PopupLinkV1 {
-  label?: string;
-  href?: string;
-}
-
-interface PopupPayloadV1 {
-  v: 1;
-  summary_lines?: string[];
-  links?: PopupLinkV1[];
-  visited?: {
-    label?: string;
-    entries?: PopupLinkV1[];
-  };
-}
-
-interface MapArgs {
-  revision: string;
-  geojson: {
-    type: "FeatureCollection";
-    features: Array<{
-      type: "Feature";
-      geometry: { type: "Point"; coordinates: [number, number] };
-      properties?: Record<string, unknown>;
-    }>;
-  };
-  height: number;
-  cluster_options?: ClusterOptionsPayload;
-  circle_marker_style?: CircleMarkerStylePayload;
-  /** Tier rgba + border/spread for ``iconCreateFunction`` (Folium cluster parity). */
-  cluster_icon_style?: ClusterIconStylePayload | Record<string, unknown>;
-  /** Injected into iframe ``<head>`` — same as Folium ``map_overlay_theme_stylesheet``. */
-  map_theme_css?: string;
-  /** Injected once — optional; leave empty. Folium ``map_popup_width_fix_script`` schedules extra shrink passes and is **not** used in the component iframe (TS owns width). */
-  map_popup_width_script?: string;
-  /** Fixed-position banner HTML (viewport = iframe), e.g. top-right ``pebird-map-banner``. */
-  banner_html?: string;
-  /** Fixed-position legend HTML, e.g. bottom-left ``pebird-map-legend``. */
-  legend_html?: string;
-  /** Camera recipe from Python `all_locations_leaflet_viewport_recipe`. */
-  viewport?: Record<string, unknown>;
-  /** Basemap keys match Python `create_map`: `default` (OSM), `google`, `carto` (CartoDB Positron). */
-  map_style?: string;
-  /** Live zoom readout (``MAP_DEBUG_SHOW_ZOOM_LEVEL`` in Python defaults). */
-  show_zoom_debug?: boolean;
-}
-
-/** Matches Folium all-locations defaults from explorer.app.streamlit.defaults. */
-const DEFAULT_CLUSTER_PAYLOAD: ClusterOptionsPayload = {
-  enabled: true,
-  max_cluster_radius: 40,
-  disable_clustering_at_zoom: 9,
-  spiderfy_on_max_zoom: false,
-  remove_outside_visible_bounds: false,
-};
-
-/** Must stay aligned with `create_map` in `explorer/presentation/map_renderer.py`. */
-type BasemapId = "default" | "google" | "carto";
-
-function normalizeBasemapId(raw: string | undefined): BasemapId {
-  const s = String(raw ?? "default").trim().toLowerCase();
-  if (s === "google" || s === "carto") {
-    return s;
-  }
-  return "default";
-}
-
-const ALL_LOCATIONS_BASEMAPS: Record<BasemapId, { url: string; opts: L.TileLayerOptions }> = {
-  default: {
-    url: "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
-    opts: {
-      maxZoom: 19,
-      attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a>',
-    },
-  },
-  google: {
-    url: "https://mt1.google.com/vt/lyrs=y&x={x}&y={y}&z={z}",
-    opts: {
-      maxZoom: 22,
-      attribution: "Google",
-    },
-  },
-  carto: {
-    url: "https://{s}.basemaps.cartocdn.com/light_all/{z}/{x}/{y}{r}.png",
-    opts: {
-      maxZoom: 20,
-      subdomains: "abcd",
-      attribution:
-        '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>',
-    },
-  },
-};
-
-function applyBasemapToMap(
-  map: L.Map,
-  mapStyleRaw: string | undefined,
-  baseTileRef: React.MutableRefObject<L.TileLayer | null>,
-): void {
-  const id = normalizeBasemapId(mapStyleRaw);
-  const spec = ALL_LOCATIONS_BASEMAPS[id];
-  if (baseTileRef.current) {
-    map.removeLayer(baseTileRef.current);
-    baseTileRef.current = null;
-  }
-  const tile = L.tileLayer(spec.url, spec.opts);
-  tile.addTo(map);
-  tile.bringToBack();
-  baseTileRef.current = tile;
-}
-
-function mergeClusterPayload(raw: ClusterOptionsPayload | undefined): ClusterOptionsPayload {
-  return { ...DEFAULT_CLUSTER_PAYLOAD, ...raw };
-}
-
-function toLeafletClusterOptions(payload: ClusterOptionsPayload): L.MarkerClusterGroupOptions {
-  return {
-    maxClusterRadius: Number(payload.max_cluster_radius) || DEFAULT_CLUSTER_PAYLOAD.max_cluster_radius!,
-    disableClusteringAtZoom:
-      Number(payload.disable_clustering_at_zoom) ||
-      DEFAULT_CLUSTER_PAYLOAD.disable_clustering_at_zoom!,
-    spiderfyOnMaxZoom: Boolean(payload.spiderfy_on_max_zoom),
-    removeOutsideVisibleBounds: Boolean(payload.remove_outside_visible_bounds),
-    chunkedLoading: true,
-  };
-}
-
-function parseClusterIconStyle(raw: unknown): ClusterIconStylePayload | null {
-  if (!raw || typeof raw !== "object") {
-    return null;
-  }
-  const o = raw as Record<string, unknown>;
-  const fills = o.fills_rgba;
-  const borders = o.borders_rgba;
-  const halos = o.halos_rgba;
-  if (!Array.isArray(fills) || !Array.isArray(borders) || !Array.isArray(halos)) {
-    return null;
-  }
-  if (fills.length !== 3 || borders.length !== 3 || halos.length !== 3) {
-    return null;
-  }
-  if (!fills.every((x) => typeof x === "string")) {
-    return null;
-  }
-  if (!borders.every((x) => typeof x === "string")) {
-    return null;
-  }
-  if (!halos.every((x) => typeof x === "string")) {
-    return null;
-  }
-  const bw = Number(o.border_width_px);
-  const spread = Number(o.halo_spread_px);
-  if (!Number.isFinite(bw) || !Number.isFinite(spread)) {
-    return null;
-  }
-  return {
-    fills_rgba: fills as string[],
-    borders_rgba: borders as string[],
-    halos_rgba: halos as string[],
-    border_width_px: bw,
-    halo_spread_px: spread,
-  };
-}
-
-function markerClusterGroupOptionsWithOptionalIconStyle(
-  payload: ClusterOptionsPayload,
-  clusterIconStyleRaw: unknown,
-): L.MarkerClusterGroupOptions {
-  const base = toLeafletClusterOptions(payload);
-  const style = parseClusterIconStyle(clusterIconStyleRaw);
-  if (!style) {
-    return base;
-  }
-  return {
-    ...base,
-    iconCreateFunction(cluster: L.MarkerCluster) {
-      const count = cluster.getChildCount();
-      const i = count < 10 ? 0 : count < 100 ? 1 : 2;
-      const html = `<div style="background-color:${style.fills_rgba[i]};border:${style.border_width_px}px solid ${style.borders_rgba[i]};box-shadow:0 0 0 ${style.halo_spread_px}px ${style.halos_rgba[i]};"><span>${count}</span></div>`;
-      const sizeClass =
-        count < 10 ? "marker-cluster-small" : count < 100 ? "marker-cluster-medium" : "marker-cluster-large";
-      return L.divIcon({
-        html,
-        className: `marker-cluster ${sizeClass}`,
-        iconSize: L.point(40, 40),
-      });
-    },
-  };
-}
-
-function applyGoToGpsViewportCamera(map: L.Map, vp: ViewportV1GoToGps): void {
-  const d = vp.epsilon_delta;
-  const b = L.latLngBounds([vp.lat - d, vp.lon - d], [vp.lat + d, vp.lon + d]);
-  map.fitBounds(b, { padding: L.point(vp.padding_px, vp.padding_px), maxZoom: vp.max_zoom, animate: false });
-}
-
-/** Red pin approximating Folium ``folium.Icon(color='red', icon='map-marker', prefix='fa')``. */
-function goToGpsMarkerIcon(): L.DivIcon {
-  return L.divIcon({
-    className: "all-locations-gps-marker",
-    html: '<span class="all-locations-gps-marker__glyph" aria-hidden="true"></span>',
-    iconSize: [30, 40],
-    iconAnchor: [15, 40],
-    popupAnchor: [0, -34],
-  });
-}
-
-/** Folium ``_apply_go_to_gps_pin_view`` marker on the map (not inside MarkerCluster). */
-function syncGoToGpsMarker(map: L.Map, viewportRaw: unknown, markerRef: React.MutableRefObject<L.Marker | null>): void {
-  if (markerRef.current) {
-    map.removeLayer(markerRef.current);
-    markerRef.current = null;
-  }
-  const vp = parseViewportV1(viewportRaw);
-  if (!vp || vp.mode !== "go_to_gps") {
-    return;
-  }
-  const m = L.marker([vp.lat, vp.lon], { icon: goToGpsMarkerIcon(), zIndexOffset: 2500 });
-  m.bindPopup(GO_TO_GPS_POPUP_HTML, POPUP_BIND_OPTIONS);
-  m.addTo(map);
-  markerRef.current = m;
-}
-
-/** Folium ``build_visit_overlay_map`` camera for All locations. */
-function applyAllLocationsViewport(map: L.Map, viewportRaw: unknown, boundsLayer: L.Layer | null): void {
-  const vp = parseViewportV1(viewportRaw);
-  const padPt = (px: number) => L.point(px, px);
-  if (!vp) {
-    if (boundsLayer && typeof (boundsLayer as L.FeatureGroup).getBounds === "function") {
-      try {
-        const b = (boundsLayer as L.FeatureGroup).getBounds();
-        if (b.isValid()) {
-          map.fitBounds(b.pad(0.12));
-        }
-      } catch {
-        map.setView([20, 0], 2);
-      }
-    } else {
-      map.setView([20, 0], 2);
-    }
-    return;
-  }
-  if (vp.mode === "go_to_gps") {
-    applyGoToGpsViewportCamera(map, vp);
-    return;
-  }
-  if (vp.mode === "center_zoom") {
-    map.setView(vp.center, vp.zoom, { animate: false });
-    return;
-  }
-  if (vp.mode === "fit_bounds") {
-    if (vp.single_point) {
-      const d = vp.epsilon_delta;
-      const b = L.latLngBounds([vp.lat - d, vp.lon - d], [vp.lat + d, vp.lon + d]);
-      map.fitBounds(b, { padding: padPt(vp.padding_px), maxZoom: vp.max_zoom, animate: false });
-      return;
-    }
-    const latlngs = vp.pairs.map((p) => L.latLng(p[0], p[1]));
-    const b = L.latLngBounds(latlngs);
-    map.fitBounds(b, { padding: padPt(vp.padding_px), maxZoom: vp.max_zoom, animate: false });
-  }
-}
-
-function escapeHtml(text: string): string {
-  const div = document.createElement("div");
-  div.textContent = text;
-  return div.innerHTML;
-}
-
-/** Keep trailing ``)`` / ``]`` from wrapping onto a line alone (typography, not lat/long-specific). */
-function preventOrphanClosingPunctuation(text: string): string {
-  return text.replace(/\s+([)\]}"'»])\s*$/u, "\u00a0$1");
-}
-
-function locationHeadingHtml(name: string, lifelistUrl: string): string {
-  const inner = escapeHtml(preventOrphanClosingPunctuation(name));
-  const hlSafe = safeHttpUrlForAnchor(lifelistUrl.trim());
-  if (hlSafe.length > 0) {
-    return `<a class="pebird-map-popup__location-heading" href="${escapeHtml(hlSafe)}" target="_blank" rel="noopener noreferrer">${inner}</a>`;
-  }
-  return `<span class="pebird-map-popup__location-heading">${inner}</span>`;
-}
-
-/** Allow only http(s) in popup anchors — blocks ``javascript:``, ``data:``, etc.  . */
-function safeHttpUrlForAnchor(raw: string): string {
-  const t = raw.trim();
-  if (!t) {
-    return "";
-  }
-  try {
-    const u = new URL(t);
-    if (u.protocol === "https:" || u.protocol === "http:") {
-      return u.href;
-    }
-  } catch {
-    /* ignore */
-  }
-  return "";
-}
-
-function parsePopupV1(raw: unknown): PopupPayloadV1 | null {
-  if (!raw || typeof raw !== "object") {
-    return null;
-  }
-  const o = raw as Record<string, unknown>;
-  if (o.v !== 1) {
-    return null;
-  }
-  const summary_lines = Array.isArray(o.summary_lines)
-    ? o.summary_lines.filter((x): x is string => typeof x === "string")
-    : undefined;
-  const linksRaw = Array.isArray(o.links) ? o.links : [];
-  const links: PopupLinkV1[] = linksRaw
-    .filter((x): x is Record<string, unknown> => !!x && typeof x === "object")
-    .map((item) => ({
-      label: typeof item.label === "string" ? item.label : "",
-      href: typeof item.href === "string" ? item.href : "",
-    }));
-  let visited: PopupPayloadV1["visited"];
-  const visRaw = o.visited;
-  if (visRaw && typeof visRaw === "object") {
-    const vo = visRaw as Record<string, unknown>;
-    const vLabel = typeof vo.label === "string" ? vo.label : "Visited:";
-    const entRaw = Array.isArray(vo.entries) ? vo.entries : [];
-    const entries: PopupLinkV1[] = entRaw
-      .filter((x): x is Record<string, unknown> => !!x && typeof x === "object")
-      .map((item) => ({
-        label: typeof item.label === "string" ? item.label : "",
-        href: typeof item.href === "string" ? item.href : "",
-      }));
-    visited = { label: vLabel, entries };
-  }
-  return { v: 1, summary_lines, links, visited };
-}
-
-/** Lifer map — structured lines from ``lifer_locations_geojson.py``. */
-interface LiferPopupLineV1 {
-  label: string;
-  date: string;
-  checklist_href: string;
-}
-
-interface LiferPopupPayloadV1 {
-  v: 1;
-  lines: LiferPopupLineV1[];
-}
-
-function parseLiferPopupV1(raw: unknown): LiferPopupPayloadV1 | null {
-  if (!raw || typeof raw !== "object") {
-    return null;
-  }
-  const o = raw as Record<string, unknown>;
-  if (o.v !== 1) {
-    return null;
-  }
-  const linesRaw = Array.isArray(o.lines) ? o.lines : [];
-  const lines: LiferPopupLineV1[] = [];
-  for (const row of linesRaw) {
-    if (!row || typeof row !== "object") {
-      continue;
-    }
-    const r = row as Record<string, unknown>;
-    lines.push({
-      label: typeof r.label === "string" ? r.label : "",
-      date: typeof r.date === "string" ? r.date : "?",
-      checklist_href: typeof r.checklist_href === "string" ? r.checklist_href : "#",
-    });
-  }
-  return { v: 1, lines };
-}
-
-/** Family map — location heading + species common-name lines (``family_locations_geojson.py``). */
-interface FamilySpeciesLineV1 {
-  name: string;
-  species_href: string;
-}
-
-interface FamilyPopupPayloadV1 {
-  v: 1;
-  species_lines: FamilySpeciesLineV1[];
-}
-
-function parseFamilyPopupV1(raw: unknown): FamilyPopupPayloadV1 | null {
-  if (!raw || typeof raw !== "object") {
-    return null;
-  }
-  const o = raw as Record<string, unknown>;
-  if (o.v !== 1) {
-    return null;
-  }
-  const linesRaw = Array.isArray(o.species_lines) ? o.species_lines : [];
-  const species_lines: FamilySpeciesLineV1[] = [];
-  for (const row of linesRaw) {
-    if (!row || typeof row !== "object") {
-      continue;
-    }
-    const r = row as Record<string, unknown>;
-    species_lines.push({
-      name: typeof r.name === "string" ? r.name : "",
-      species_href: typeof r.species_href === "string" ? r.species_href : "",
-    });
-  }
-  return { v: 1, species_lines };
-}
-
-/** Species map — structured sections from ``species_locations_geojson.py``. */
-interface SpeciesObservationV1 {
-  datetime_label: string;
-  checklist_href: string;
-  observed_count: string;
-  media_href: string;
-}
-
-interface SpeciesSectionV1 {
-  common_name: string;
-  observation_count: number;
-  open_by_default: boolean;
-  observations: SpeciesObservationV1[];
-}
-
-interface SpeciesVisitsBlockV1 {
-  summary_label: string;
-  entries: PopupLinkV1[];
-  open_by_default: boolean;
-}
-
-interface SpeciesPopupPayloadV1 {
-  v: 1;
-  location_heading_margin_px?: number;
-  species_sections: SpeciesSectionV1[];
-  visits: SpeciesVisitsBlockV1;
-}
-
-function parseSpeciesPopupV1(raw: unknown): SpeciesPopupPayloadV1 | null {
-  if (!raw || typeof raw !== "object") {
-    return null;
-  }
-  const o = raw as Record<string, unknown>;
-  if (o.v !== 1) {
-    return null;
-  }
-  const sectionsRaw = Array.isArray(o.species_sections) ? o.species_sections : [];
-  const species_sections: SpeciesSectionV1[] = [];
-  for (const sec of sectionsRaw) {
-    if (!sec || typeof sec !== "object") {
-      continue;
-    }
-    const s = sec as Record<string, unknown>;
-    const obsRaw = Array.isArray(s.observations) ? s.observations : [];
-    const observations: SpeciesObservationV1[] = [];
-    for (const row of obsRaw) {
-      if (!row || typeof row !== "object") {
-        continue;
-      }
-      const r = row as Record<string, unknown>;
-      observations.push({
-        datetime_label: typeof r.datetime_label === "string" ? r.datetime_label : "",
-        checklist_href: typeof r.checklist_href === "string" ? r.checklist_href : "",
-        observed_count: typeof r.observed_count === "string" ? r.observed_count : "",
-        media_href: typeof r.media_href === "string" ? r.media_href : "",
-      });
-    }
-    species_sections.push({
-      common_name: typeof s.common_name === "string" ? s.common_name : "",
-      observation_count:
-        typeof s.observation_count === "number" && Number.isFinite(s.observation_count)
-          ? s.observation_count
-          : observations.length,
-      open_by_default: s.open_by_default === true,
-      observations,
-    });
-  }
-  const visRaw = o.visits;
-  if (!visRaw || typeof visRaw !== "object") {
-    return null;
-  }
-  const vo = visRaw as Record<string, unknown>;
-  const entRaw = Array.isArray(vo.entries) ? vo.entries : [];
-  const entries: PopupLinkV1[] = entRaw
-    .filter((x): x is Record<string, unknown> => !!x && typeof x === "object")
-    .map((item) => ({
-      label: typeof item.label === "string" ? item.label : "",
-      href: typeof item.href === "string" ? item.href : "",
-    }));
-  const visits: SpeciesVisitsBlockV1 = {
-    summary_label: typeof vo.summary_label === "string" ? vo.summary_label : "Visited:",
-    entries,
-    open_by_default: vo.open_by_default === true,
-  };
-  const margin =
-    typeof o.location_heading_margin_px === "number" && Number.isFinite(o.location_heading_margin_px)
-      ? o.location_heading_margin_px
-      : 6;
-  return { v: 1, location_heading_margin_px: margin, species_sections, visits };
-}
-
-/** Species-map matching pin — mirrors ``assemble_species_map_location_popup_html``. */
-function popupHtmlSpeciesLayout(
-  name: string,
-  lifelistUrl: string,
-  payload: SpeciesPopupPayloadV1,
-): string {
-  const margin =
-    typeof payload.location_heading_margin_px === "number" &&
-    Number.isFinite(payload.location_heading_margin_px)
-      ? payload.location_heading_margin_px
-      : 6;
-  const locHeading = locationHeadingHtml(name, lifelistUrl);
-
-  const sectionParts: string[] = [];
-  for (const sec of payload.species_sections) {
-    const openAttr = sec.open_by_default ? " open" : "";
-    const summaryLabel = `${sec.common_name}: (${sec.observation_count})`;
-    const obsLines: string[] = [];
-    for (const obs of sec.observations) {
-      const hrefSafe = safeHttpUrlForAnchor(obs.checklist_href.trim());
-      const dt = escapeHtml(obs.datetime_label.trim() || "—");
-      const count = escapeHtml(obs.observed_count.trim());
-      let line =
-        hrefSafe.length > 0
-          ? `<a href="${escapeHtml(hrefSafe)}" target="_blank" rel="noopener noreferrer">${dt}</a>`
-          : `<span>${dt}</span>`;
-      line += ` <span class="pebird-map-popup__obs-count">(Observed: ${count})</span>`;
-      const mediaSafe = safeHttpUrlForAnchor(obs.media_href.trim());
-      if (mediaSafe) {
-        line += ` <a class="pebird-map-popup__media-link" href="${escapeHtml(mediaSafe)}" target="_blank" rel="noopener noreferrer" title="media">↗</a>`;
-      }
-      obsLines.push(`<div class="pebird-map-popup__obs-line">${line}</div>`);
-    }
-    sectionParts.push(
-      `<details class="pebird-map-popup__species-seen"${openAttr}>` +
-        `<summary class="pebird-map-popup__section-label">${escapeHtml(summaryLabel)}</summary>` +
-        `<div class="pebird-map-popup__obs-list">${obsLines.join("")}</div>` +
-        `</details>`,
-    );
-  }
-
-  const visitAnchors: string[] = [];
-  for (const e of payload.visits.entries) {
-    const href = e.href?.trim() ?? "";
-    const linkLabel = e.label?.trim() || href;
-    if (href) {
-      const hrefSafe = safeHttpUrlForAnchor(href);
-      if (hrefSafe) {
-        visitAnchors.push(
-          `<a href="${escapeHtml(hrefSafe)}" target="_blank" rel="noopener noreferrer">${escapeHtml(linkLabel)}</a>`,
-        );
-      } else {
-        visitAnchors.push(`<span class="pebird-map-popup__visit-link-text">${escapeHtml(linkLabel)}</span>`);
-      }
-    }
-  }
-  const visitsOpen = payload.visits.open_by_default ? " open" : "";
-  const visitsBlock =
-    `<details class="pebird-map-popup__all-visits"${visitsOpen}>` +
-    `<summary class="pebird-map-popup__section-label">${escapeHtml(payload.visits.summary_label)}</summary>` +
-    `<div class="pebird-map-popup__visit-list-inner">${visitAnchors.join("<br>")}</div>` +
-    `</details>`;
-
-  return (
-    `<div class="pebird-map-popup popup-scroll-wrapper" style="position:relative;">` +
-    `<div class="pebird-map-popup__heading-row" style="margin-bottom:${margin}px;">${locHeading}</div>` +
-    `<div class="pebird-map-popup__scroll" style="max-height:300px;overflow-y:auto;">` +
-    sectionParts.join("") +
-    visitsBlock +
-    `</div></div>`
-  );
-}
-
-/** Per-pin circle resolved in Python (``circle_pin``) for Lifer vs subspecies colours. */
-interface CirclePinPayload {
-  stroke_hex?: string;
-  fill_hex?: string;
-  radius_px?: number;
-  stroke_weight?: number;
-  fill_opacity?: number;
-  stroke_opacity?: number;
-}
-
-function parseCirclePin(raw: unknown): CirclePinPayload | null {
-  if (!raw || typeof raw !== "object") {
-    return null;
-  }
-  const o = raw as Record<string, unknown>;
-  return {
-    stroke_hex: typeof o.stroke_hex === "string" ? o.stroke_hex : undefined,
-    fill_hex: typeof o.fill_hex === "string" ? o.fill_hex : undefined,
-    radius_px: typeof o.radius_px === "number" && Number.isFinite(o.radius_px) ? o.radius_px : undefined,
-    stroke_weight:
-      typeof o.stroke_weight === "number" && Number.isFinite(o.stroke_weight) ? o.stroke_weight : undefined,
-    fill_opacity:
-      typeof o.fill_opacity === "number" && Number.isFinite(o.fill_opacity) ? o.fill_opacity : undefined,
-    stroke_opacity:
-      typeof o.stroke_opacity === "number" && Number.isFinite(o.stroke_opacity)
-        ? o.stroke_opacity
-        : undefined,
-  };
-}
-
-function resolvedCircleStylesFromPinPayload(pin: CirclePinPayload): Pick<
-  L.CircleMarkerOptions,
-  "radius" | "weight" | "color" | "fillColor" | "fillOpacity"
-> {
-  const okHex = (s: string | undefined, fb: string) =>
-    typeof s === "string" && /^#[0-9a-fA-F]{6}$/.test(s) ? s : fb;
-  const fillHex = okHex(pin.fill_hex, "#3388ff");
-  const strokeHex = okHex(pin.stroke_hex, "#1c2630");
-  const radius =
-    typeof pin.radius_px === "number" && Number.isFinite(pin.radius_px) && pin.radius_px > 0
-      ? pin.radius_px
-      : 7;
-  const weight =
-    typeof pin.stroke_weight === "number" && Number.isFinite(pin.stroke_weight) && pin.stroke_weight >= 1
-      ? pin.stroke_weight
-      : 1;
-  let fillOp = 0.88;
-  if (typeof pin.fill_opacity === "number" && Number.isFinite(pin.fill_opacity)) {
-    fillOp = Math.min(1, Math.max(0, pin.fill_opacity));
-  }
-  return {
-    radius,
-    weight,
-    color: strokeHex,
-    fillColor: fillHex,
-    fillOpacity: fillOp,
-  };
-}
-
-/** Family composition popup — mirrors ``format_family_location_popup_html``. */
-function popupHtmlFamilyLayout(name: string, lifelistUrl: string, payload: FamilyPopupPayloadV1): string {
-  const margin = POPUP_LOCATION_HEADING_MARGIN_PX;
-  const locHeading = locationHeadingHtml(name, lifelistUrl);
-  const lineParts: string[] = [];
-  for (const ln of payload.species_lines) {
-    const n = ln.name.trim();
-    if (!n) {
-      continue;
-    }
-    const hrefSafe = safeHttpUrlForAnchor(ln.species_href.trim());
-    if (hrefSafe) {
-      lineParts.push(
-        `<div class="pebird-map-popup__species-line"><a href="${escapeHtml(hrefSafe)}" target="_blank" rel="noopener noreferrer">${escapeHtml(n)}</a></div>`,
-      );
-    } else {
-      lineParts.push(`<div class="pebird-map-popup__species-line">${escapeHtml(n)}</div>`);
-    }
-  }
-  const bodyHtml =
-    lineParts.length > 0
-      ? lineParts.join("")
-      : '<div class="pebird-map-popup__summary-line">No species lines</div>';
-  return (
-    `<div class="pebird-map-popup popup-scroll-wrapper" style="position:relative;">` +
-    `<div class="pebird-map-popup__heading-row" style="margin-bottom:${margin}px;">${locHeading}</div>` +
-    `<div class="pebird-map-popup__scroll" style="max-height:300px;overflow-y:auto;">${bodyHtml}</div>` +
-    `</div>`
-  );
-}
-
-/** Lifer popup: location heading + species checklist lines (parity with ``format_lifer_popup_lines``). */
-function popupHtmlLiferLayout(name: string, lifelistUrl: string, payload: LiferPopupPayloadV1): string {
-  const margin = POPUP_LOCATION_HEADING_MARGIN_PX;
-  const locHeading = locationHeadingHtml(name, lifelistUrl);
-  const lineParts: string[] = [];
-  for (let i = 0; i < payload.lines.length; i += 1) {
-    const ln = payload.lines[i];
-    const prefix = i > 0 ? "<br>" : "";
-    const hrefRaw = ln.checklist_href.trim() || "#";
-    const hrefSafe = safeHttpUrlForAnchor(hrefRaw);
-    const label = ln.label.trim() || "—";
-    const dateStr = ln.date.trim() || "?";
-    if (hrefSafe) {
-      lineParts.push(
-        `${prefix}<a href="${escapeHtml(hrefSafe)}" target="_blank" rel="noopener noreferrer">${escapeHtml(
-          label,
-        )} : ${escapeHtml(dateStr)}</a>`,
-      );
-    } else {
-      lineParts.push(`${prefix}<span>${escapeHtml(label)} : ${escapeHtml(dateStr)}</span>`);
-    }
-  }
-  const inner = lineParts.join("");
-  return (
-    `<div class="pebird-map-popup popup-scroll-wrapper" style="position:relative;">` +
-    `<div class="pebird-map-popup__heading-row" style="margin-bottom:${margin}px;">${locHeading}</div>` +
-    `<div class="pebird-map-popup__scroll" style="max-height:300px;overflow-y:auto;">` +
-    `<div class="pebird-map-popup__visited-block">` +
-    `<div class="pebird-map-popup__visit-dates">${inner}</div>` +
-    `</div></div></div>`
-  );
-}
-
-/** Classic All locations card — DOM mirrors ``assemble_location_popup_html`` / ``LocationPopupModel`` (``map_popup_models``). */
-function popupHtmlVisitedLayout(
-  name: string,
-  lifelistUrl: string,
-  visited: NonNullable<PopupPayloadV1["visited"]>,
-): string {
-  const label = visited.label?.trim() || "Visited:";
-  const entries = visited.entries ?? [];
-  const margin = POPUP_LOCATION_HEADING_MARGIN_PX;
-  const locHeading = locationHeadingHtml(name, lifelistUrl);
-
-  const visitAnchors: string[] = [];
-  for (const e of entries) {
-    const href = e.href?.trim() ?? "";
-    const linkLabel = e.label?.trim() || href;
-    if (href) {
-      const hrefSafe = safeHttpUrlForAnchor(href);
-      if (hrefSafe) {
-        visitAnchors.push(
-          `<a href="${escapeHtml(hrefSafe)}" target="_blank" rel="noopener noreferrer">${escapeHtml(linkLabel)}</a>`,
-        );
-      } else {
-        visitAnchors.push(`<span class="pebird-map-popup__visit-link-text">${escapeHtml(linkLabel)}</span>`);
-      }
-    }
-  }
-  /** Mirrors Folium ``build_visit_info_html``: ``<br>`` between *inline* checklist links — not ``display:block`` anchors. */
-  const visitInner = visitAnchors.join("<br>");
-
-  return (
-    `<div class="pebird-map-popup popup-scroll-wrapper" style="position:relative;">` +
-    `<div class="pebird-map-popup__heading-row" style="margin-bottom:${margin}px;">${locHeading}</div>` +
-    `<div class="pebird-map-popup__scroll" style="max-height:300px;overflow-y:auto;">` +
-    `<div class="pebird-map-popup__visited-block">` +
-    `<div class="pebird-map-popup__section-label">${escapeHtml(label)}</div>` +
-    `<div class="pebird-map-popup__visit-dates">${visitInner}</div>` +
-    `</div>` +
-    `</div></div>`
-  );
-}
-
-/** Single Leaflet popup layout for structured `popup_v1` (+ legacy fallback). */
-function popupHtmlFromFeatureProps(props: Record<string, unknown> | undefined): string {
-  const name = String(props?.name ?? "Location");
-  const lifelistUrl = String(props?.lifelist_url ?? "");
-  const liferPop = parseLiferPopupV1(props?.lifer_popup_v1);
-  if (liferPop) {
-    return popupHtmlLiferLayout(name, lifelistUrl, liferPop);
-  }
-  const familyPop = parseFamilyPopupV1(props?.family_popup_v1);
-  if (familyPop) {
-    return popupHtmlFamilyLayout(name, lifelistUrl, familyPop);
-  }
-  const speciesPop = parseSpeciesPopupV1(props?.species_popup_v1);
-  if (speciesPop) {
-    return popupHtmlSpeciesLayout(name, lifelistUrl, speciesPop);
-  }
-  const popup = parsePopupV1(props?.popup_v1);
-  if (popup?.visited) {
-    return popupHtmlVisitedLayout(name, lifelistUrl, popup.visited);
-  }
-  if (popup) {
-    const margin = POPUP_LOCATION_HEADING_MARGIN_PX;
-    const locHeading = locationHeadingHtml(name, lifelistUrl);
-    let html =
-      `<div class="pebird-map-popup">` +
-      `<div class="pebird-map-popup__heading-row" style="margin-bottom:${margin}px;">${locHeading}</div>`;
-    for (const line of popup.summary_lines ?? []) {
-      html += `<span class="pebird-map-popup__summary-line">${escapeHtml(line)}</span>`;
-    }
-    for (const link of popup.links ?? []) {
-      const href = link.href?.trim() ?? "";
-      const label = link.label?.trim() || "Link";
-      if (href) {
-        const hrefSafe = safeHttpUrlForAnchor(href);
-        if (hrefSafe) {
-          html += `<span class="pebird-map-popup__summary-line"><a href="${escapeHtml(hrefSafe)}" target="_blank" rel="noopener noreferrer">${escapeHtml(
-            label,
-          )}</a></span>`;
-        } else {
-          html += `<span class="pebird-map-popup__summary-line">${escapeHtml(label)}</span>`;
-        }
-      }
-    }
-    html += "</div>";
-    return html;
-  }
-  const visits = props?.visit_checklists;
-  const url = String(props?.lifelist_url ?? "").trim();
-  const margin = POPUP_LOCATION_HEADING_MARGIN_PX;
-  const locHeading = locationHeadingHtml(name, url);
-  let legacy =
-    `<div class="pebird-map-popup">` +
-    `<div class="pebird-map-popup__heading-row" style="margin-bottom:${margin}px;">${locHeading}</div>`;
-  if (visits != null && visits !== "") {
-    legacy += `<span class="pebird-map-popup__summary-line">Checklists: ${escapeHtml(String(visits))}</span>`;
-  }
-  legacy += "</div>";
-  return legacy;
-}
-
-function isHex6(s: string | undefined): boolean {
-  return typeof s === "string" && /^#[0-9a-fA-F]{6}$/.test(s);
-}
-
-/** Folium-equivalent CircleMarker options from Python or legacy GeoJSON ``colour``. */
-function resolvedCircleStyles(
-  cm: CircleMarkerStylePayload | undefined,
-  featureColour: string | undefined,
-): Pick<L.CircleMarkerOptions, "radius" | "weight" | "color" | "fillColor" | "fillOpacity"> {
-  const fillHex = isHex6(cm?.fill_hex)
-    ? cm!.fill_hex!
-    : isHex6(featureColour)
-      ? featureColour!
-      : "#3388ff";
-  const strokeHex = isHex6(cm?.stroke_hex) ? cm!.stroke_hex! : "#1c2630";
-  const radius =
-    typeof cm?.radius_px === "number" && Number.isFinite(cm.radius_px) && cm.radius_px > 0
-      ? cm.radius_px
-      : 7;
-  const weight =
-    typeof cm?.stroke_weight === "number" &&
-    Number.isFinite(cm.stroke_weight) &&
-    cm.stroke_weight >= 1
-      ? cm.stroke_weight
-      : 1;
-  let fillOp = 0.88;
-  if (typeof cm?.fill_opacity === "number" && Number.isFinite(cm.fill_opacity)) {
-    fillOp = Math.min(1, Math.max(0, cm.fill_opacity));
-  }
-  return {
-    radius,
-    weight,
-    color: strokeHex,
-    fillColor: fillHex,
-    fillOpacity: fillOp,
-  };
-}
-
-const STYLE_ID = "pebird-map-overlay-theme";
-const POPUP_WIDTH_SCRIPT_ID = "pebird-map-popup-width-fix";
-
-/** Join inner CSS from every ``<style>...</style>`` block (Python concatenates popup + banner/legend sheets). */
-function extractAllStyleInnerCss(html: string): string {
-  const s = html.trim();
-  if (!s) {
-    return "";
-  }
-  const parts: string[] = [];
-  const re = /<style[^>]*>([\s\S]*?)<\/style>/gi;
-  let m: RegExpExecArray | null;
-  while ((m = re.exec(s)) !== null) {
-    parts.push(m[1].trim());
-  }
-  if (parts.length > 0) {
-    return parts.join("\n");
-  }
-  /* Fallback: legacy single-block strip */
-  return s
-    .replace(/^\s*<style[^>]*>\s*/i, "")
-    .replace(/\s*<\/style>\s*$/i, "")
-    .trim();
-}
-
-/** Inner JS from ``<script>...</script>`` (single block from Python). */
-function extractScriptInnerJs(html: string): string {
-  const s = html.trim();
-  if (!s) {
-    return "";
-  }
-  const m = /<script[^>]*>([\s\S]*?)<\/script>/i.exec(s);
-  return m ? m[1].trim() : s.replace(/^\s*<script[^>]*>\s*/i, "").replace(/\s*<\/script>\s*$/i, "").trim();
-}
-
-/** Folium ``add_zoom_level_debug_overlay`` / ``_ZoomLevelDebugOverlay`` in ``map_renderer.py``. */
-function syncZoomLevelDebugOverlay(
-  map: L.Map,
-  enabled: boolean,
-  controlRef: React.MutableRefObject<L.Control | null>,
-  onZoomRef: React.MutableRefObject<(() => void) | null>,
-): void {
-  const existing = controlRef.current;
-  if (existing) {
-    map.removeControl(existing);
-    controlRef.current = null;
-  }
-  const onZoom = onZoomRef.current;
-  if (onZoom) {
-    map.off("zoom", onZoom);
-    map.off("zoomend", onZoom);
-    onZoomRef.current = null;
-  }
-  if (!enabled) {
-    return;
-  }
-  const div = L.DomUtil.create("div", "ebird-zoom-debug-overlay");
-  div.style.cssText = [
-    "background:rgba(255,255,255,0.92)",
-    "border:1px solid #1f6f54",
-    "padding:4px 8px",
-    "font:12px/1.25 ui-monospace, SFMono-Regular, Menlo, monospace",
-    "border-radius:4px",
-    "box-shadow:0 1px 3px rgba(0,0,0,0.2)",
-    "min-width:7ch",
-    "text-align:right",
-    "z-index:1001",
-  ].join(";");
-  const ZoomDebugControl = L.Control.extend({
-    onAdd: () => div,
-  });
-  const ctrl = new ZoomDebugControl({ position: "bottomright" });
-  ctrl.addTo(map);
-  const update = () => {
-    div.textContent = `zoom: ${map.getZoom()}`;
-  };
-  map.on("zoom", update);
-  map.on("zoomend", update);
-  update();
-  controlRef.current = ctrl;
-  onZoomRef.current = update;
-}
-
-function injectHeadFragments(mapThemeCss: string, mapPopupWidthScript: string): void {
-  const css = (mapThemeCss ?? "").trim();
-  if (css) {
-    let styleEl = document.getElementById(STYLE_ID) as HTMLStyleElement | null;
-    if (!styleEl) {
-      styleEl = document.createElement("style");
-      styleEl.id = STYLE_ID;
-      document.head.appendChild(styleEl);
-    }
-    const inner = extractAllStyleInnerCss(css);
-    styleEl.textContent = inner;
-  }
-  const scr = (mapPopupWidthScript ?? "").trim();
-  if (scr && !document.getElementById(POPUP_WIDTH_SCRIPT_ID)) {
-    const inner = extractScriptInnerJs(scr);
-    const s = document.createElement("script");
-    s.id = POPUP_WIDTH_SCRIPT_ID;
-    s.textContent = inner;
-    document.head.appendChild(s);
   }
 }
